+++ src/components/Todolist.tsx
@@ -2,7 +2,7 @@ export default function TodoList() {
   return (
     <div className="flex flex-col w-full">
       <div className="divider divider-start divider-neutral">
-      <input type="text" placeholder="Title" className="input input-ghost w-full max-w-xs" />
+        <p className="text-black">hoge</p>
       </div>
     </div>
   );
